refactor: MentoringApplicationItem에서 상태 이모티콘 로직을 MentoringApplicationStatus 컴포넌트로 이동 #192
